Add ItemComponent prop
