+++ portal/src/UserProfileAttributesList.module.scss
@@ -1,7 +1,3 @@
-.fieldName {
-  font-weight: 600;
-}
-
 .dragAndDrop {
   cursor: pointer;
 }
